VarianceForecaster/pyproject.toml
[tool.poetry]
name = "varianceforecaster"
version = "0.1.0"
description = ""
authors = ["Дмитрий Цитман <dimazzik@gmail.com>"]
readme = "README.md"
packages = [
    { include = "vf_app" }
]

[tool.poetry.group.dev.dependencies]
httpx = "^0.28.1"

[tool.mypy]
plugins = []
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.poetry.dependencies]
python = ">=3.12"
pwlf = "^2.5.1"
numpy = "^2.3.1"
scikit-learn = "^1.7.0"
pytest = "^8.4.1"
matplotlib = "^3.10.3"
xgboost = "^3.0.2"
fastapi = "^0.116.1"
jinja2 = "^3.1.6"
mypy = "^1.17.0"
uvicorn = "^0.35.0"
isort = "^6.0.1"
passlib = "^1.7.4"
pyjwt = "^2.10.1"
sqlalchemy = "^2.0.41"
bcrypt = "^4.3.0"
python-multipart = "^0.0.20"
